nimble/host: Fix format string in ble_gap_log_duration

+++ nimble/host/src/ble_gap.c
@@ -320,7 +320,7 @@ ble_gap_log_duration(int32_t duration_ms)
     if (duration_ms == BLE_HS_FOREVER) {
         BLE_HS_LOG(INFO, "duration=forever");
     } else {
-        BLE_HS_LOG(INFO, "duration=%dms", duration_ms);
+        BLE_HS_LOG(INFO, "duration=%dms", (int)duration_ms);
     }
 }
 #endif
